Made publish class lightier
