+++ src/Kmd.Logic.Gateway.Automation/PublishFile/Product.cs
@@ -2,7 +2,7 @@
 
 namespace Kmd.Logic.Gateway.Automation.PublishFile
 {
-    internal class Product
+    public class Product
     {
         public string Name { get; set; }
